+++ src/main/java/com/sangraj/carrental/entity/Car.java
@@ -3,6 +3,8 @@ package com.sangraj.carrental.entity;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Entity
 @Data
 @Table(name = "cars")
@@ -19,7 +21,7 @@ public class Car {
     private String fuelType;
     private int seats;
     @Column(name = "\"pricePerDay\"")
-    private double pricePerDay;
+    private BigDecimal pricePerDay;
     private String location;
     private Integer quantity;
     private String imageUrl;
